server/requirements.txt
redis
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.8.5
djangorestframework==3.11.0
django-restapi==0.1.3
python-dateutil